Files moved to about directory
