+++ src/pages/about/TeamDetails.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { AboutCard, Container } from '.';
-import { man1 } from '../images';
+import { AboutCard, Container } from '../../components';
+import { man1 } from '../../images';
 
 export default function TeamDetails({ teamName, img }) {
 	return (
